Use the mail utility rather than Cron emails.
The motivation for this change is to get slightly nicer emails (than the
standard cron output) when the sanity check fails.  Although I think the
result might be nicer, it's obviously made the command we run a lot more
complicated.

I've left the MAILTO line in there so that we receive emails if the
command fails in unexpected ways.

I've installed the mail command line tool through the mailutils package
on our server.